Add latest explorer to mesh
Since commands will be applied in a sequence the explorer that will
receive the subsequent commands should be well-known

The Mars Explorer defines a mesh boundary then land one or more explorers
applying a sequence of commands that drives a particular exploration

Storing on system state allows to retrieve the right explorer to send
commands. Its not a problem replace the explorer right after by a new
one since the commands were sent.